Move test helpers to the test suite .go file
After moving the podnic.go tests to podnic_test.go put the tests helpers
at the network_suite_test.go so they are in a common place.
